requirements.txt
arrow==0.13.1
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.7.1
binaryornot==0.4.4
certifi==2019.3.9
cffi==1.12.2
constantly==15.1.0
cookiecutter==1.6.0
cryptography==2.6.1
cssselect==1.0.3
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jinja2-time==0.2.0
lxml==4.3.3
numpy==1.16.2
pandas==0.24.2
parsel==1.5.1
Pillow==6.0.0
poyo==0.4.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pytesseract==0.2.6
python-dateutil==2.8.0
pytz==2019.1
queuelib==1.5.0
Scrapy==1.6.0
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.8
TowelStuff==0.1.dev0
Twisted==19.2.0
urllib3==1.24.1
w3lib==1.20.0
webencodings==0.5.1
whichcraft==0.5.2
zope.interface==4.6.0
